changed constant names and moved command groups to strategies package
