+++ src/org/usfirst/frc4904/robot/RobotMap.java
@@ -56,14 +56,12 @@ public class RobotMap {
 		public static class Serial {
 			public static final int navX = 0;
 			public static final int ballInnie = 2;
-			// public static final int trayInnie = 3; ----- PROJECT CURRENTLY PAUSED
 			public static final int flywheelLeftMotor = 4;
 		}
 		
 		public static class CAN {
 			public static final int leftEncoder = 0x602;
 			public static final int rightEncoder = 0x603;
-			// public static final int trayEncoder = 0x604; ---- not being used anymore
 			public static final int flywheelEncoder = 0x605;
 			public static final int elevatorEncoder = 0x606;
 			public static final int lidarTurnEncoder = 0x607;
