+++ src/org/usfirst/frc4904/robot/autonomous/strategies/AutonVomitRed.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.autonomous;
+package org.usfirst.frc4904.robot.autonomous.strategies;
 
 public class AutonVomitRed extends AutonVomit {
